Refactor and add non-typed routes example
