+++ src/main/kotlin/task11/ktor/task01.kt
@@ -36,9 +36,8 @@ fun main() {
                 }
             }
 
-            index()
-            cocktails(dao)
-            ingredients(dao)
+            cocktails()
+//            cocktailsWithLocations(dao)
         }
     }.start(wait = true)
 }
